Layer qt refactor (#90)
Refactor code base to put layers qt files in layers folder, and also pass non-UI peer class to the constructer of the UI class which adds its as a property so all of the connections from the UI class to the non-UI class happens within the UI class now. The non-UI classes now use the vispy event system to inform the UI classes when they need to update.