Add repr_png method to Viewer
I was getting kind of annoyed that notebooks would only display
"napari.Viewer object at <0xGIBBERISH>" at the end of cells
calling napari. This instead makes a png and displays it.

The disadvantage is that users might be confused and think that
this is the interactive napari viewer. But, on the whole, I think
it's a pretty cool way to save notebook state!