Initial experimental performance monitoring
If NAPARI_PERFMON is defined we use our own QApplication
that times events and sends them to our new PerfTimers
class.

If NAPARI_PERFMON_TRACE_PATH is defined them PerfTimers
will write the timers to a chrome://tracing support log file.

NAPARI_PERFMON also creates a new dock widget QtPerformance
which is very much a WIP. It shows the average UpdateRequest time
which the main paint/draw event that hangs if we are doing IO
during the draw. It also has a text window where it "lots" any
"slow" events (UpdateRequest or otherwise).