New perfmon that works with Jupyter (#3)
New convert_app_for_timing works with Jupyter

The old approach was restricted to gui_qt which is only
used when running napari from the command line.

Now we have 2 approaches. We still do gui_qt for the
console app. But in Viewer we call convert_app_for_timing
and this works for Jupyter.