Add event to string name mapping for PyQt5
In PySide2 the str() method gives you a nice string name like
'PySide2.QtCore.QEvent.Type.KeyPress' But in PyQt5 it return
a string/integer like '6' which is annoying.

This creates a lookup table on startup. This should be 100% find and
perform fine. So we use it for both PyQt5 and PySide2 to be consisting
so we work the same in both.