Performance monitoring (#1262)
* Initial experimental performance monitoring

If NAPARI_PERFMON is defined we use our own QApplication
that times events and sends them to our new PerfTimers
class.

If NAPARI_PERFMON_TRACE_PATH is defined them PerfTimers
will write the timers to a chrome://tracing support log file.

NAPARI_PERFMON also creates a new dock widget QtPerformance
which is very much a WIP. It shows the average UpdateRequest time
which the main paint/draw event that hangs if we are doing IO
during the draw. It also has a text window where it "lots" any
"slow" events (UpdateRequest or otherwise).

* Cleanup, remove unused code, remove print

Create LONG_EVENT_MS class var
Remove print and replace with comment.

* Remove stale comment

* Better comments

* Improve comments

* Set max/min not needed

* perf_timers and tracing comments

Big comment on the lack of full nesting
support for PerfTimers

* Define a perf_counter_ns() to support Python 3.6

* Cleanup QtPerformance, add "uptime"

Add "draw time" and "slow events' labels.
Add lots of comments.
Extract method: _get_timer_info

* Better comment describing the layout

* Add perf_timer context object

* Big iteration of perfmon stuff, put in own utils/perf folder (#1)

* A lot of cleanup/improvement to perf stuff
* Move USE_PERFMON check to event_loop.py
* Split things into mutiple files
* New Debug menu and trace file recording menu options.

* Remove stray method

* use perf.USE_PERFOMON instead of env var

* Add a combo box to set threashold (#2)

* Fix decorator

* Need category, placeholder for now

* New perfmon that works with Jupyter (#3)

New convert_app_for_timing works with Jupyter

The old approach was restricted to gui_qt which is only
used when running napari from the command line.

Now we have 2 approaches. We still do gui_qt for the
console app. But in Viewer we call convert_app_for_timing
and this works for Jupyter.

* Fix comment

* Better classname

* Better comment about QApplication

* Add event to string name mapping for PyQt5

In PySide2 the str() method gives you a nice string name like
'PySide2.QtCore.QEvent.Type.KeyPress' But in PyQt5 it return
a string/integer like '6' which is annoying.

This creates a lookup table on startup. This should be 100% find and
perform fine. So we use it for both PyQt5 and PySide2 to be consisting
so we work the same in both.

* Better comment on convert_app_for_timing call

* Edits for readability. Plus add sys.argv to the QApplication()

* Merge in perf_func/perf_func_name from other branch.

Most of the time we can just use perf_func that uses the function name
by default. Most of the time that's what we want.

* Talley review: moves/fixed Parameters for __init__

* Talley review: allow docking to the right

I just picked "bottom" early on before I knew what was going to be in the
widget. Works fine on the right so yes allowing that here.

* Talley review: use env var directly and late import

Avoid importing perf at all unless we are really using it.

* Better comment

* Talley review: remove allowed_areas entirely