Factor out ImageSlice and ImageView from Image (#1343)
* Factor out ImageSlice and ImageView

Previously:
    _data : the full data, all slices
    _data_view : viewable current slice
    _data_raw : raw current slice
    _data_thumbnail : viewable current thumbnail
    N/A             : raw current thumbnail

    Image class updates raw->view:
        self._data_raw = np.clip(image, 0, 1)
        self._data_view = self._raw_to_displayed(self._data_raw)

Now:
    _data : same as before
    _slice.image.view : viewable current slice
    _slice.image.raw : raw current slice
    _slice.thumbnail.view : viewable current thumbnail
    _slice.thumbnail.raw : raw current thumbnail

    ImageSlice class updates raw->view
        self._slice.image.raw = np.clip(image, 0, 1)

* Move factored out method to creat a cleaner diff (hopefully)

* Remove: not part of this refactoring

* Extract 2 files for ImageSlice and ImageView

* Better comments

* Add comment

* Comments, type annotations

* Remove temp comment

* Change to image_slice since slice is reserved (just in commment)

* Keep Image._data_view and Image.data_raw for now

Some tests use Image.data_raw but tons of tests and non-tests
use Image._data_view. Maybe we can move away from it at some
point but for now we need it.

* Remove line that we strongly believe was doing nothing

Nick on Zulip:
all the tests pass on master without it suggesting that it can be removed, i do have vague memories of why it was there, potentially related to passing "multiscale" labels and needing an initialization to be correct, but looking through the code now I'm really pretty confident it isn't needed anymore. At this point if removing it introduces a bug that isn't caught by our current tests I'd want to add a test and properly fix the bug instead of leaving this "hack" in so to speak

* Fix labels layers: allow it to pass in a raw->view converter

This a more powerful and cleaner setup than having _raw_to_displayed
be a method of Image that gets overridden by Labels.

Since we pass in the converter, Labels or any other layer can pass
in any function they want, they don't have to just define one single
override. Also it remove the pointless "null" version of
_raw_to_displayed that Image used it have.

Minor points but still better.

* Put _raw_to_displayed back in Image class

Tests forbid Image.__init__ from having any parameter that's
not returned by _get_state, so we can't pass in the converter
to Image.__init__.

Not a big deal this now works as before where the converter
is hardcoded based on the layer class. Less flexible but fine.

* Remove now unused param to __init__

* Fix doc string

* kn42 review: fix doc strings

* use doctest format in ImageView docstr

* doctest for ImageView also