New way to handle blank + placeholder
We can't set _slice.image.view to be the placeholder because people expect
it to immediately have the shape of the set image.

Now we create a "full size" blank image immediately, so we have the right
shape. And we create the placeholder off-to-the-side.

In `VispyImageLayer._on_data_change` we now access
the new `Image.displayed_slice` instead of `Image._data_view`

The new `Image.displayed_slice` will be the placholder if we are still
loaded, while `Image.data_view` will be the full size blank.

This works for all tests EXCEPT when we try to access the actual data
immediately after setting a new image. With async it's not there yet.
TBD what to do there. Do we have to break the API and require an
explicit load()? Or do we sense the bytes are being access and
force the load?