Remove drawing the placeholder to fix some tests
The placeholder with text on it was not really kosher, because it meant we
were drawing a 1024x1024 image when that was not the real image size.
This led to some tests breaking, for example test_big_2D_image because
it expected the drawn image was downsampled but it wasn't, since we
were drawing 1024x1024 and not 20,000x10.

Swapping in the placehold was kind of a hack, and it was not possible to
use the same trick for 3D images. So pulling it out makes sense.

For real use we'll need some type of vispy text overlay or just a
Qt dialog box, or something else. Something that's totally separate and
on top of the data we are drawing.