Create QtChunkReceiver to fix threading/dependency
We didn't want ChunkLoader to depend on Qt at all. So we have it use an
EmitterGroup to signal. Then we create a new QtChunkReceiver which can
connect to ChunkLoader's event. But it MIGHT get singled in a worker
thread, could be either one. So it signals itself with Qt Signal/Slot which
switched things over to the GUI thread. Then we can call into the
model/layer code in the GUI thread.