DOC: update param names to match function signature.
This attempt to match – in some non ambiguous case parameter names in
the docstring of functions to actually match the name in the signature.

This so far only works when there is a single misspelling and all the
other parameters are documented.

This also try respect a couple of other things with heuristics, to match
numpydocs:
  - Parameters with no type given should not have a ` : `
  - if there is a `:` without spaces assume it is still the type
  separator.