Revent image.py to original, add AsyncImage
So that users not using NAPARI_ASYNC see as few changes
as possible, let's try leaving the origina Image class in
place. And only use the new AsyncImage class if
NAPARI_ASYNC is set.

It doesn't seem right to destabalize users who do not
opt-in to the new async class.