Redo AsyncImage from scratch to be a lot simpler
We spent a long time in this PR getting single-scale images working.
Pretty late in the process we started making the node invisible
while we were loading. Then we got mult-scale working.

By this time AsyncImage had a lot of changes that were
redundant or could be simplifed. So pretty much started over
relying more heavily on the "invisible" trick. Turns out a lot
of the intial changes made getting single-scale working
weeks ago were not really needed.

So this new version is a lot simpler and a much smaller
change, and both single-scale and multi-scale work.