Add new QtChunkReceiver
The QtChunkReceiver listens for the `chunk_loaded` event from the
ChunkLoader that it will receive in a worker thread, and it calls
the Layer's `chunk_loaded` method in the GUI thread.

We don't want the workers directly calling into model/layer code. The
QtChunkReceiver is Qt because Qt's slots/signals are the only way
we know to have a worker thread post a message to the GUI thread.