Fix random "missing layer" bug in DelayQueue.
Basically a re-write. It addition to fixing the bug it
no longer wakes up periodically. It always sleeps
until the next request is due, or if there are no
requests it politely waits on an event.

So the run-time behavior should now be pretty much
perfect, exactly what we want.