Add visply Visual.loaded boolean
This fixes test_viewer.py::test_add_layer. Before we
did not set_data() if the layer was invisible. But
test_add_layer adds insivisble layers and exceptects
then to be set correctly.

So no we use the Visual.loaded flag instead of visible,
because only we set the loaded on. So the test
still works.