Fix #1631 by creating a new slice every time (#1641)
* Fix 1631 by creating a new slice *every* time

The idea that we needed to preserve a slice even when _set_view_slice()
was based on async work that I think is moot now. It was based on an
update-loop type of bug where _set_view_slice() was called redundantly.

With that fixed it's safe to just create an entire new slice. This will work
even with the async branch.

This is much cleaner anyway so _set_view_slice() means "create the slice".

* add screenshot test labels painting

* Update .pre-commit-config.yaml

Co-authored-by: Nicholas Sofroniew <sofroniewn@gmail.com>