async-2.6: Move async into experimental  (#1659)
* Fix async after _new_empty_slice() change

We added _new_empty_slice() in master to fix labels. But
this actually broke async because in async we were calling
refresh when the chunks was loaded which emptied the slice.

For now we just called events.set_data() and _update_thumbnail()
manually. Will need more thought.

* Prep for most to experimental

Create ImageSliceData, SyncImageLoader and AsyncImageLoader
So that we can have a clear line between master/sync and experimental/async

* Fix multiscale typo

* Move perf into components.experimental.perf

* Fix method name

* Fix use_async test!

* Blank __init__.py needed for pytest

* Move QtChunkReceiver into experimental

* Slim down comments and normalize them.

* Fully guard it with NAPARI_ASYNC

* Only create global if NAPARI_ASYNC

* Disable test temporarily

* Add _use_async

* Slim down comment

* Move experimental import down

* Add experimental AsyncImageSliceData

* Do not import experimental at all unless async

* if statement not needed

* Guard experimental code against import

* Derive from ImageLoader

Not really needed, but to be consistent with
AsyncImageSliceData and ImageSliceData.

* Comments, rename chunk_request -> request

* Check for pytest

pytest imports everything,

* Detect with sys.modules?

* Remove tests for now

These tests made sure we don't import async code
unless NAPARI_ASYNC is defined. But they are failing
in CI (pytest?). It must import everything regradless
of whether it is referenced.

Might be a way around this but remove it for now.

* More doc string

* Rename some Async classes to Chunked

Although the feature toggle is NAPARI_ASYNC the
ChunkLoader does both sync and async loading.

So better to call these Chunked not Async.

* Docs and return bool instead of data

Returning the data was just not necessary.

* Fix name 'Chunked'

* Doc string and fix return value

* Blank since others are blang

* Use forward reference for return type

* Remove rgb test for now

This breaks something with sync. This test was added
to avoid getting some mysterious error when your
rgb flag was wrong. If that error comes up again
we can try to figure out a way to flag it earlier
before the load finishes.