async-5: Timing Infrastructure (#1678)
* Initial async-2: thread pool only

* Factor out a generic QtGuiEvent object

* Prevent error on shutdown

* Remove unused wait method

* Comments fix async/sync check.

* Better comment

* More doc string

* Comments

* Comments

* Disconnect to eliminate crash on exit

* Fix close() call

* Comments, logging

* doc string and rename member

* Improve doc strings

* Remove types because people prefer...

That we write same types over and over
and over.

* Tweak comment

* Simplify using all()

* Not using load_seconds yet so remove it

* Better comment

* Layer return type

* Better comment

* Streamline

* Rename event for clarity

* Add long comments to explain.

* Review: import futures differently

* Fix typo in comment

* Improve wording

* Add comment about logging and f-strings

* Vispy changes and ChunkCache

* Depend on cachetools

* comments

* Add new AsyncConfig

* Start of new perf_timer which yields

* Add _PYTEST_RAISE from #1510

This is very useful for debugging with
vs code, so you it will break where
the exception happens.

* Add visply Visual.loaded boolean

This fixes test_viewer.py::test_add_layer. Before we
did not set_data() if the layer was invisible. But
test_add_layer adds insivisble layers and exceptects
then to be set correctly.

So no we use the Visual.loaded flag instead of visible,
because only we set the loaded on. So the test
still works.

* Put loaded here instead of base Visual class

Since base Visuals is not vendored.
And since this is hopefully temporary
doesn't make sense to vendor it
just for this.

* Add cachetools

* Default to loaded for non-async case

* Check node.loaded not load.visible

That's why we added node.loaded!
Nodes can be insivible for non-async reasons.

* Use self.layer not self.node

Used to be "visible" but since it "loaded" we might as well
ask the layer

* Removed loaded from vendored image/volume

Put checks in VispyImageLayer only per nick.
Avoids making changes to the vendored files.

* Fix chunk timing

Python default args are only evaluated
once when the function is defined!
Unlike C++ wherer they are evaluated
for each function call.

* Add loaded flag from async-3

* Fix update loop bug with async

If load is in progress, don't try to
load again.

* Fix comment

* Vendor cachetools

Roughly following PIP's policy:
https://pip.pypa.io/en/stable/development/vendoring-policy/

* Not working yet

* Add np.all to fix test

* pre-hash the key, simpler and faster

* Add _vendor cachetools

* exclude vendor from Black format

* Add _vendor to manifest

* Try different manifest syntax

* try graft?

* Introduce block_timer

block_timer works even if perfmon is disabled

* Merge from master

* add _vendor/experimental/cachetools

* Add _vendor/experimental/cachetools

* Merge async-2.7 (vendor cachetools)

* Make it empty again

* Remove conftest.py change from PR

I will create a separate PR for this since it's unrelated.

* Make cachetools import a bit nicer

You could argue this is not necessary, but I think it
from "vendor.experimental.cachetools" is the clearest
verison of what we are doing there.

* Add comments

* Fix CI error: we don't need _vendor tests

* Fix CI error: we don't need _vendor tests

No point in having cachetools tests in our tree.

* Merge async-3 into here

* Move to experimental

* Comments

* Uncomment - useless

* Comment - it's not a class

* Comments

* Update comment to latest file locations

* Comments