Fix ValueError in Layer.coordinates in fringe cases
With Octree discovered moving the mouse while a layer is
loading can cause this ValueError. This reason is strange,
the call to get_max_texture_sizes() would create a temporary
Canvas, which would lead to vispy processing Qt Events
which meant mousemove() events were handled when
the layer was only partiall loaded.