Add location and elapsed_ms
elapsed_ms is so we can time chunks "end to end" instead of just timing the load portion. This is not fully setup yet, but we've been
logging it.

This led us to discover that in-progress
loads no longer need by the octree
were not being cancelled. But we've
add support for that now, still debugging.