Start work on touch events (pinch to zoom and co).
This does need to use a fork of of vizpy,I'll send pull request to vispy
progressively and try to make it work both on stable and master of
vispy.

Currently need https://github.com/vispy/vispy/pull/1968

I'll try to thread that into the canvas to various dimensions but still
try to wrap my head around napari codebase.