DOC: Rename Section to conform to numpydocs (Return->Returns)
The numpydoc style guide
(https://numpydoc.readthedocs.io/en/latest/format.html) indicate that
the name of the section is "Returns", not "Return", and in one case the
content of the section is dedented for uniformity.

Note that "Returns" section are supposed to use the same formatting as
the "Parameters" section and have the type of what they return as a
definition list; but that's a longer endeavour.