DOC: autoreformat all the docstrings.
This is based on my vélin prototype which parses docstings using
numpydoc and reformat them to follow the numpydoc format to the t.

It is of course a prototype so should be checked carefully and not
blindly trusted. I'll send another pull request to try to make it part
of CI and precomit.