Use vectorized indexing for xarray arrays in labels (#2191)
* Use vectorized indexing for xarray arrays in labels

It's important to use DataArrays to index, otherwise the memory
usage can explode as it will do orthogonal indexing instead of
vectorized indexing. See
http://xarray.pydata.org/en/stable/indexing.html#vectorized-indexing
for details.

* add a test for painting with xarray

The failure mode is that xarray attempts to index a huge chunk of the
array. e.g. shape (1024, 1024, 1024, 1024) which eats up memory and thus
slows down the test. There does not seem to be a pytest mark for memory
usage so use timing as a proxy