Generic Button to toggle state.
To work better with the action manager we do want a more generic way to
represent buttons that toggle between two states; currently this is only
done in two places: QtNDisplayButton and QtGridViewButton whcih are
checkboxes and not Button; thus they do not have the same interface.

Instead of subclassing checkbox and target the qss on the new class;
create generic QtBistateButton that sets the mode for qss targetting and
has a well defined set of parameters.

There is an increase in number of indirection; so a small increase of
complexity; but this is likely reusable for other button ; for example
the toggle console button is a push button and coudl reflect wether the
console is shown or not.