Fix some translation formatting.
I believe the call to format need to me outside the call for trans,
it make no sens  otherwise, say for example.

    trans._("Open file {file}".format(file=file))

.format() is called before trans, and there is no way for the
translation to provide translation for all the filenames,

Though in the opposite
    trans._("Open file {file}").format(file=file)

then the translation can provide the alternative:

    "Ouvrir le fichier {file}"

Which is then formatted.

This seem to be the same for KEY_SYMBOLS, the symbol depends on the
platform, and the user plaform, this I believe need to be translated
before formatting. If we want to translate thing like "Control",
"Space", "Shift" then those need to be done directly in the KEY_SYMBOLS,
like :

KEY_SYMBOLS = {
    'Shift': trans._('Shift'),
    ...
}

But not the other way around.