Add sticky notification panel.
This adds a sticky notification panel that keep reference to all the
previous notifications. It does not allow (yet) to access all the
buttons like show traceback, as this may not always make sens once the
action effect of a potential button has been triggered once.

This uses list item, which appearance is similar to Plugins items.