[Prototype] Centralized shortcut management
This is a prototype to try to synchronize shortcut management, naming
and button tooltip as well as help with future configurability.

This adds a central location on which you register a callback (with a
name) and use it as a proxy to bind callback on buttons... etc.

This central command/action manager thus has enough knowledge to update
the buttons tooltips when the keyboard shortcuts are changed, and could
read from settings to update those.

There is probably nothing to keep from this; but I was thiking to use
this as a discussion platform for ppwadhwa's work

forgot to bind callback

sync

Try to make the Shortcut manager works with the Keymap provider...

rename instance to action_manager

sync

some command -> actions renaming

cleanup

stuff

bind more keys

moar short

Wrong instance passed through.

remove commented code

cleanup

sync

flash buttons

and rebased