Centralized shortcut management
This is a prototype to try to synchronize shortcut management, naming
and button tooltip as well as help with future configurability.

This adds a central location on which you register a callback (with a
name) and use it as a proxy to bind callback on buttons... etc.

This central command/action manager thus has enough knowledge to update
the buttons tooltips when the keyboard shortcuts are changed, and could
read from settings to update those.

This allows global manipulation of (some) shortcuts; in order to accommodate
user preferences, localization and ensure that buttons and various items have
proper descriptions. This also layout the foundation to record actions.

When using the action manager, binding buttons, qaction, shortcut ...etc is
capable of inserting itself, to record when actions are triggered. To show this,
using this PR, issue for example a transpose (Ctrl-T) action, and you will see
that the transpose button flashes (which is a nice touch when demoing).

Rebase of d795fcabb

try to fix deprecation

take some comments into accounts

recording

dont trigger on change for demo

rename _animation to _flash_animation

typo

rename keymappable to keymapprovider

was 1b02356c2