Centralized shortcut management
This try to consolidate shortcut management, naming
and button tooltip and menu, as well as help with future configurability.
This is purposefully incomplete to keep the change of functionality
relatively small without changing too much of the internal API in
particular.

The KeymapProvider approach conflict with the possibility of putting all
QAction with shortcuts in the menu.

If a QAction with a shortcut is added to a menu, its shortcut is bound
with the main window, so it will not trigger at the KeymapProvider
level. For layers this is a problem as currently the KeyMapProvider
instance will be the one that receive the shortcut and pass self to the
callback. I believe this is fixable, but I'd like to keep it out of
scope of this PRs.

We also want to keep the KeymapProvider to bind actions (for now), so
the action manager need to do some dance and store the shortcuts and
buttons separately from the Keymap provider.

For translation and configurability it also make sens from a API point
of view to not register the shortcut at the same time as the action.
Indeed, the default shortcut that could be registered with the action
may conflict with local shortcuts, or be overridden with user
preferences. I thus think the API should _encourage_ action makes the
recognize that the two are different. Note that it is easy to make a
context manager that let you register a function with a shortcut at the
same time if you wish to for plugins authors.

With respect to creating buttons, one does not need to directly bind the
actions, but can ask the action_manger to do so, this will let the
action manager update the button tooltips with the currently active
shortcuts.

The central action manager also have a notion of context. Currently this
is only a dict which is directly manipulated but will be updated in teh
future. This lets us manipulate the global state an actin will see and
the parameters it will receive.

For example an action which callback has the signature f(viewer), will
be called with the value of context['viewer'] as the first parameter.

This let us define actions outside of existing classes (which can be
useful for plugins) and will let us migrate some action from
KeyMapProvider to global function for example taking the `active_layer`
as a parameters. I believe this will make recording much simpler; as
then the flow of action goes through the action manager and be
serialised and replayed; instead of getting actions directly handled by
instances in which case tracking what action has been done can be hard.

Remove any logic related to flashing, or recording. See
10693f0fa388b2fa1ada3aa03712135e291675ff for a commit with flashing and
recording logic.