Extract shortcut into their own object.
Spawn out of Action manager; Shortcut formatting is not limited to
swapping the right symbols for Control, Alt and alike.
The join character is not the same on platform.

It will also be useful for interoperability with QT which uses the +
symbol for joining values.

I've implemented the __str__ method to be platform specific, but I'm not
100% certain, so I'm happy with changing the impl

This also make basic checks that the shortcut looks alright when
instantiating it.