Extend the action manager to support multiple shortcuts
And add a few actions.

When multiple shortcuts are bound they will now be show in tooltips
comma separated.

unbind_shortcut unbind all the existing shortcuts for a given action,
that seem the most reasonable thing to do, though maybe it should
pluralized ? Though as most actions have only one shortcut, I'm not sure
that make sens.

This also use the action manager for a couple more actions, in
particular on the shapes layers the move to front and move to back
now have an assigned actions, even if they do not have shortcuts.
This will let users add keybindings to those if they wish to.