Do not store default values in preference files.
When a value match the default value from napari in settings; do not
store it.

This is useful as this let us change the default values across napari
versions and  leads to much smaller settings.

This should also address #2847 which was driving me crazy as changinf
default_values of shortcuts had no effects.