Prototype : handle "hold XX to" with action manager.
I initially went the direction of having isgeneratorfunction in
`register_action` but realise we do not really want to have two action
with different names for the generator and the non generator. Typically
pan_zoom.

So I went the route of allowing an alternate implementation for when we
register an action.
 - the normal one
 - the one for "hold to XXX".

Thus there is now 2 bind_shortcut: bind_shortcut and bind_hold_shortcut.

Both interact with the tooltips so that a button that allow to toggle to
pan zoom, with get (XXX or YYY or hold ZZZZ) as a tooltip.

I'm also thinking that maybe there is room from Qt Buttons  that have
effects only while pressed. Definitely not pan to zoom as you use the
mouse for something else, but maybe YAGNI.