Tifffile compress' kwargs deprecated. Update to compression.
This is since version 2021 6 6, I've also sent a patch to tifffile
itself to set the proper stacklevel for the deprecation warnings, see
`cgohlke/tifffile#89`.

I'm comparing the version numbers and LBYL instead of EAFP as I find it
much easier to cleanup and find dead code due to change in version
number when those are explicit.