Include the message in notification REPR.
There are a couple times when test failures in pytest show
ErrorNotification problem in tests; unfortunately the repr does not show
the error message. This is dues to Events repr only including attributes
that are properties and Notification being a subclass of these.

I'm not sure why the repr only include properties and not other kind of
attributes – maybe to not include methods ? Or if attributes are
themselves to big ? So I turn the message attribute into a property.

Closes #2873