Alternate "Hold xxx for action" + action invert.
Unlike #2854 which let user define action that are either functions, or
generator this allows to register an "invert" for an action.

The invert takes the same parameter as the original action plus a state
object (what is returned from an action).

In the action manager as we are still relying on KeymapProvider,
we turn this into a generator and bind that on a key (so we can get hold
space to pan zoom for example), but we could store the intermediate
state and keep a stack of actions we can undo.

We _could_ likely also let users simply register a generator with both
side of the yield being the action and its invert,

and then either bind the generator for "hold to xxx", or lambda:
next(gen()) to only do the forward step, though with this we have no way
of recording a stack and have undo/redo; unless we keep the generator
instances arround, which I'm not a fan of.

Thoughts on wether you have preference for this approach or the one in
gh-2854 are welcome.