handle multiple (qt)viewer in multiple windows.
I'm trying to introduce the notion of active viewer, it's not perfect but help
for #2977, so it does not for #2897, though discussing with Juan, we are not
sure it's worth fixing when both are in the same window.

Which viewer should be targeted by actions when there are many is
inherently ambiguous with the concept of action and menu - especially on mac
where there is a single menu, while there are often one menu per window on other
platform.

This also update the toggle console visibility action even if regardless, we
don't support multiple console and only the first one open will show the prompt.