Shape Mouse refactor
Working on #2907 I realized the logic was a bit complex and could use
some generalisation, especially if we want to add more customisability.

This uses dictionaries and mapping to simply declare which callback
should be active with which modes, which should be easier to make
configurable, and also I believe easier to understand and less
error-prone (you only need to change what need to be active instead of
thinking about what need to be added and removed on mode change).

I've made the values class attributes, which mean that you should be
able to change the configuration at runtime by poking directly at the
class, and it should work as long as you don't change config of current
mode.

The only weird thing if maybe the _REV_SHAPE_HELP, I build the
dictionaries backward before of reversing it to only have a single
instance of each translated string as I'm not too sure how it deals with
duplication.

This also allows layer to register action on double clicks.

This does the same as for mouse-press and release, bur register double
click. Note that according to vispy not all platform which are supported
y it have native support for double click and therefore this is
"emulated" double click.

It is common to think that when one bind double click, and user double
clicks then the single click action is not triggered. Unfortunately as
vispy is performance based this is not the case as we want to react on
keydown for actions and if we want a double click we would need to at
least delay single actions by the maximum time a double click takes.

Thus double click is _in addition to_ the action performed by single
click.

Fortunately this does not seem to be too much of a problem in most case.
  - if there is no single click action bound we are all good.
  - double click can also often "undo" the work of the single click.

This also uses the double click infrastructure to allow for finishing a
shape by double click (the same way Inkscape, Illustrator do it).

cleanup and document

Add Tests

Just change the explicit call to finish by a double click event.

forgotten merge