Shape Mouse refactor
Working on #2907 I realized the logic was a bit complex and could use
some generalisation, especially if we want to add more customisability.

This uses dictionaries and mapping to simply declare which callback
should be active with which modes, which should be easier to make
configurable, and also I believe easier to understand and less
error-prone (you only need to change what need to be active instead of
thinking about what need to be added and removed on mode change).

I've made the values class attributes, which mean that you should be
able to change the configuration at runtime by poking directly at the
class, and it should work as long as you don't change config of current
mode.

The only weird thing if maybe the _REV_SHAPE_HELP, I build the
dictionaries backward before of reversing it to only have a single
instance of each translated string as I'm not too sure how it deals with
duplication.