Split windows pyside/pyqt into own GitHub action check
I find it relatively confusing in the github action UI to find issues on
windows that happen only on Pyqt/pyside, I regularly forget that the
windows action run both, and I scoll to find the error for a minute or
so before figuring out that the I need to expand the other section.

I also find that having from the GitHub UI indication of which
variant(s) failed is a good way to guess wether we have a random failure
or an actual issue.

I'm hoping this may also reduce the "time to CI results" as we can now
run pytest/pyqt in parallel.