Attempt to cache tox virtualenv.
It looks like the venv creation is what takes about 4min/gh-action.
Hopefully caching will be faster.

The cache should be trashed if setup.cfg or tox.ini changes.

I cache the whole .tox/ dir but my understanding is that it's ok as the
cache keys will be under different names so no need to cache the
specific venv.

in order to flush the cache size we need a hack like:

https://stackoverflow.com/questions/60942067/get-current-date-and-time-in-github-workflows