Speedup one of the slowest test.
Creating a 100B random values takes most of the time in this test (5 to
6 sec), which is unnecessary. calc_data_range return [0,255] if the type
is uint8 anyway w/o looking at the data.

I guess we could reduce even further, but maybe those large values are
to test that da actually chunks the array, but at that point this is
quasi-instantaneous anyway.