Refactor the Point, Label and Shape Layer Mode logic.
The Shape was already refactor, this does the same refactoring in the
Points and Label Layer, and move the common logic to a helper in the
base Layer.

This should make it slightly easier to make new layers with modes
and different mouse callbacks depending on which mode is active.

This make it also a tiny bit less error-prone to edit what is active in
which mode.

There is still a bit of repetition to keep things readable, and I erred
on the side of expliciteness (having a no_op callback) instead of of
checking whether a callback is present.

There is a hint of complexity with cursor because of square brush, which
is deprecated and should have been removed.