Catch and prune connections to stale Qt callbacks in events.py (#3193)
In working on #3184, it became clear that the biggest hurdle to reuse of our evented objects (after closing a Qt window), is the fact that python wrappers for "deleted" QtObjects stick around long after the QObject is deleted, causing the classic `RuntimeError: wrapped C/C++ object has been deleted` problem if an `EventEmitter` tries to call a method that belongs to a dead object.

`events.py` already _does_ try to catch whether a connected callback is no longer useful: it stores weakrefs to the object that owns the callback method, and if that object has been garbage collected, [simply refuses to call the callback and deletes it](https://github.com/napari/napari/blob/52c449a675ebf90540188c5a05b46638aca7bbac/napari/utils/events/event.py#L560-L565).  It seems to me that we should also have that behavior for connected Qt objects...  but it's a bit tricker since their python wrappers don't get garbage collected.  

This PR catches a call to a dead Qt object and quietly deletes the callback, just as it would for a dead non-Qt python object.

Nick suggested that we could also add a warning when the callback is deleted, but we can still do that later if required.