Found while revieweing #3178
This typo:

https://github.com/napari/napari/pull/3178/files#r689815054

Should have triggered a traceback somewhere, but did not.

I tracked that down to the IPython Verbose traceback formatter,
when trying to inspect some frames to display the local variables,
it triggers a SyntaxError when accessing a frame attribute.

More precisely:

    > frame_info.variables_in_executing_piece

raises a SyntaxError (i'm guessing this goes deeper into stackdata).

I'll add a workaround in IPython to not raise but show "Exception trying
to inspect frame. No more locals available." when we can't see locals,
but in the meantime this at least let napari users see that something is
wrong.

I'm also asserting the loops values are correct