Mitigate #3228
This remove the crash/exception that makes things behave a bit better,
but not quite.

I think we should consolidate the _is_moving and _moving_coordinates
into a single item to not have inconsistency when manipulating those
values.

It's still unclear to me how this behave exactly, but it seem the loginc
is a anyway a bit wrong as pressing shift before drawing forces 45º,
but not after drawing the first point.

note that with lines we likely want to allow all multiple of 45º and not only
45º+k*90º.